Fixing seed script
